#871 refactor(ui-review): share browser observations
